+++ Ablaze/src/Scene/Components/Motion/Transform.h
@@ -54,7 +54,7 @@ namespace Ablaze
 		void Rotate(float angle, const Maths::Vector3f& axis, Space space = Space::World, Angle angleType = Angle::Radians);
 		void Rotate(const Maths::Quaternion& quaternion);
 
-		Maths::Matrix4d ToMatrix() const;
+		Maths::Matrix4f ToMatrix() const;
 		String ToString() const override;
 
 		Component* Clone() const override;
